requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.10.1
pyjwkest==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-taiga==1.0.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
